refactor: split code by logical domains
